Don't use SECT_4K mode on the winbond w25q256 flash
MYIR started populating this part on SOMs, and it does not work with UBIFS,
because UBIFS requires a minimum erase block of 16kb. However, the flash
part still supported 64k erase commands (for backwards compatibility I
presume), and removing the SECT_4K causes the m25p80 to use this command,
and the mtd device to report a LEB of 64K, just like the old flash.